+++ foxes/models/point_models/__init__.py
@@ -5,3 +5,4 @@ Point models.
 from .wake_deltas import WakeDeltas
 from .set_uniform_data import SetUniformData
 from .tke2ti import TKE2TI
+from .ustar2ti import Ustar2TI