+++ CHANGELOG.md
@@ -702,5 +702,7 @@ This major version introduces the concept of `Engines` which handle the chunking
   - Now three directories are defined: `work_dir`, `input_dir`, `output_dir`. If not defined, falling back to `work_dir`, defaulting to cwd.
 - Inputs:
   - Removing `SliceDataNC`, since now `FieldDataNC` can handle data without height dependency
+- Models:
+  - New point model `Ustar2TI`, estimates TI from friction velocity
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.2.3](https://github.com/FraunhoferIWES/foxes/commits/v1.2.3)